classes/packaged-staging.bbclass: * special case glibc-intermediate till we solve the shlib problem * use ipkg with a package name instead of a file * rebuild the package-index each time * make filename checks less strict

+++ classes/packaged-staging.bbclass
@@ -25,11 +25,12 @@
 DEPLOY_DIR_PSTAGE 	= "${DEPLOY_DIR}/pstage" 
 
 PSTAGE_BUILD_CMD        = "${IPKGBUILDCMD}"
-PSTAGE_INSTALL_CMD      = "ipkg-cl install -f ${DEPLOY_DIR_PSTAGE}/ipkg.conf -force-depends -o "
+PSTAGE_INSTALL_CMD      = "ipkg-cl install -f ${DEPLOY_DIR_PSTAGE}/ipkg.conf -o "
+PSTAGE_UPDATE_CMD	= "ipkg-cl update -f ${DEPLOY_DIR_PSTAGE}/ipkg.conf -o "
 PSTAGE_PKGNAME 		= "staging-${PN}_${PV}-${PR}_${PACKAGE_ARCH}.ipk"
 
 SPAWNFILE 		= "${STAGING_DIR}/pkgmaps/${P}-${PR}.spawn"
-SPAWNIPK		= "${DEPLOY_DIR_IPK}/${spawn}_${PV}-${PR}_${PACKAGE_ARCH}.ipk"
+SPAWNIPK                = ${spawn}
 
 do_clean_append() {
         """clear the build and temp directories"""
@@ -45,7 +46,7 @@ do_stage_prepend() {
 #detect aborted staging attempts
 if [ -e ${TMPDIR}/moved-staging ]; then
         oenote "Detected a moved staging, moving it back"
-	${TMPDIR}/pstage ${STAGING_DIR} && rm ${TMPDIR}/moved-staging
+	rm -rf ${STAGING_DIR} && ${TMPDIR}/pstage ${STAGING_DIR} && rm ${TMPDIR}/moved-staging
 fi
 
 if [ -e ${STAGING_DIR} ]; then
@@ -80,6 +81,10 @@ ipkgarchs="all any noarch ${TARGET_ARCH} ${IPKG_ARCHS} ${IPKG_EXTRA_ARCHS} ${MAC
       echo "arch $arch $priority" >> ${DEPLOY_DIR_PSTAGE}/ipkg.conf
       priority=$(expr $priority + 5)
     done
+echo "src oe file:${DEPLOY_DIR_IPK}" >> ${DEPLOY_DIR_PSTAGE}/ipkg.conf 
+export OLD_PWD=`pwd`
+cd ${DEPLOY_DIR_IPK} && ipkg-make-index -p Packages . ; cd ${OLD_PWD}
+${PSTAGE_UPDATE_CMD} ${STAGING_DIR}
 
 #blacklist packages poking in staging *and* cross
 if [ ${PN} != "linux-libc-headers" ] ; then
@@ -87,10 +92,10 @@ if [ ${PN} != "linux-libc-headers" ] ; then
 	if [ -e ${SPAWNFILE} ]; then
         	oenote "List of spawned packages found: ${P}.spawn"
         	for spawn in `cat ${SPAWNFILE} | grep -v ${PN}-locale` ; do \
-			if [ -e ${SPAWNIPK} ]; then
-               	 		${PSTAGE_INSTALL_CMD} ${STAGING_DIR} ${SPAWNIPK}          
+			if [ -e ${DEPLOY_DIR_IPK}/${spawn}* ]; then
+               	 		${PSTAGE_INSTALL_CMD} ${STAGING_DIR} ${spawn}          
 			else
-				oenote "${SPAWNIPK} not found, probably empty package"
+				oenote "${spawn} not found, probably empty package"
 			fi
         	done
         	exit 0
@@ -99,45 +104,53 @@ if [ ${PN} != "linux-libc-headers" ] ; then
 	fi
 fi #if ${PN}
 
-if [ -e ${DEPLOY_DIR_PSTAGE}/${PSTAGE_PKGNAME} ]; then
-	oenote "Staging stuff already packaged, using that instead"
-	${PSTAGE_INSTALL_CMD} ${STAGING_DIR}  ${DEPLOY_DIR_PSTAGE}/${PSTAGE_PKGNAME}
-	exit 0      
-fi
 
-touch ${TMPDIR}/moved-staging
-mv ${STAGING_DIR} ${TMPDIR}/pstage
+if [ ${PN} != "glibc-intermediate" ] ; then
 
-mkdir -p ${STAGING_BINDIR}
-mkdir -p ${STAGING_LIBDIR}
-mkdir -p ${STAGING_INCDIR}
-mkdir -p ${STAGING_DATADIR}/aclocal
+	if [ -e ${DEPLOY_DIR_PSTAGE}/${PSTAGE_PKGNAME} ]; then
+		oenote "Staging stuff already packaged, using that instead"
+		${PSTAGE_INSTALL_CMD} ${STAGING_DIR}  ${DEPLOY_DIR_PSTAGE}/${PSTAGE_PKGNAME}
+		exit 0      
+	fi
 
+	touch ${TMPDIR}/moved-staging
+	mv ${STAGING_DIR} ${TMPDIR}/pstage
+
+	mkdir -p ${STAGING_BINDIR}
+	mkdir -p ${STAGING_LIBDIR}
+	mkdir -p ${STAGING_INCDIR}
+	mkdir -p ${STAGING_DATADIR}/aclocal
+else	
+	oenote "Glibc-intermediate detected!"
+fi #if !glibc intermediate
 }
 
 do_stage_append() {
 
-mkdir -p ${DEPLOY_DIR_PSTAGE}
-mkdir -p ${STAGING_DIR}/CONTROL
-echo "Package: staging-${PN}"           >  ${STAGING_DIR}/CONTROL/control
-echo "Version: ${PV}-${PR}"             >> ${STAGING_DIR}/CONTROL/control
-echo "Description: ${DESCRIPTION}"      >> ${STAGING_DIR}/CONTROL/control
-echo "Section: ${SECTION}"              >> ${STAGING_DIR}/CONTROL/control
-echo "Priority: Optional"               >> ${STAGING_DIR}/CONTROL/control
-echo "Maintainer: ${MAINTAINER}"        >> ${STAGING_DIR}/CONTROL/control
-echo "Architecture: ${PACKAGE_ARCH}"    >> ${STAGING_DIR}/CONTROL/control
-echo "Source: ${SRC_URI}"               >> ${STAGING_DIR}/CONTROL/control
+if [ ${PN} != "glibc-intermediate" ] ; then
+	mkdir -p ${DEPLOY_DIR_PSTAGE}
+	mkdir -p ${STAGING_DIR}/CONTROL
+	echo "Package: staging-${PN}"           >  ${STAGING_DIR}/CONTROL/control
+	echo "Version: ${PV}-${PR}"             >> ${STAGING_DIR}/CONTROL/control
+	echo "Description: ${DESCRIPTION}"      >> ${STAGING_DIR}/CONTROL/control
+	echo "Section: ${SECTION}"              >> ${STAGING_DIR}/CONTROL/control
+	echo "Priority: Optional"               >> ${STAGING_DIR}/CONTROL/control
+	echo "Maintainer: ${MAINTAINER}"        >> ${STAGING_DIR}/CONTROL/control
+	echo "Architecture: ${PACKAGE_ARCH}"    >> ${STAGING_DIR}/CONTROL/control
+	echo "Source: ${SRC_URI}"               >> ${STAGING_DIR}/CONTROL/control
 
-mkdir -p ${DEPLOY_DIR_PSTAGE}
+	mkdir -p ${DEPLOY_DIR_PSTAGE}
 
-${PSTAGE_BUILD_CMD} ${STAGING_DIR} ${DEPLOY_DIR_PSTAGE}
+	${PSTAGE_BUILD_CMD} ${STAGING_DIR} ${DEPLOY_DIR_PSTAGE}
 
-rm -rf ${STAGING_DIR}
-#move back stagingdir so we can install packages   
-mv ${TMPDIR}/pstage ${STAGING_DIR}
-rm ${TMPDIR}/moved-staging
+	rm -rf ${STAGING_DIR}
+	#move back stagingdir so we can install packages   
+	mv ${TMPDIR}/pstage ${STAGING_DIR}
+	rm ${TMPDIR}/moved-staging
 
-${PSTAGE_INSTALL_CMD} ${STAGING_DIR}  ${DEPLOY_DIR_PSTAGE}/${PSTAGE_PKGNAME}
-
+	${PSTAGE_INSTALL_CMD} ${STAGING_DIR}  ${DEPLOY_DIR_PSTAGE}/${PSTAGE_PKGNAME}
+else
+	oenote "Glibc-intermediate detected (again)"
+fi #if !glibc-intermediate
 }
